requirements.txt
google-generativeai>=0.7.2
duckduckgo-search>=6.2.6
python-dotenv>=1.0.1

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0
